chore(webapp): Fix to reorganised code in subfolders
